Bug fixes Generated numbers now match Mersenne Twister reference page http://www.math.sci.hiroshima-u.ac.jp/~m-mat/MT/MT2002/emt19937ar.html

+++ rand/randmt.h
@@ -23,6 +23,6 @@ RAND_MT *mtrand_new(unsigned long seed);
 void mtrand_dispose(RAND_MT *mt);
 
 /* Get random number. */
-int mtrand_get(RAND_MT *mt);
+unsigned long mtrand_get(RAND_MT *mt);
 
 #endif /* Z_RAND_MT */
